FrontEnd responsive, order /mymeetups by date

All routes tested

+++ frontend/src/pages/Details/index.js
@@ -20,30 +20,34 @@ export default function Details({ match }) {
     getMeetup();
   }, [match]);
 
+  function handleEdit() {
+    return meetup.past
+      ? toast.error('You cannot edit a meetup that already happened')
+      : history.push(`/meetups/${meetup.id}/edit`);
+  }
+
   async function handleCancel() {
+    if (meetup.past)
+      return toast.error('You cannot cancel a meetup that already happened');
     await api.delete(`meetups/${meetup.id}`);
-    toast.warn(`You have canceled ${meetup.name}!`);
-    history.push('/dashboard');
+    toast.success(`You have canceled ${meetup.name}!`);
+    return history.push('/dashboard');
   }
 
   return (
     <Container>
       <header>
         <h1>{meetup.name}</h1>
-        <span>
-          <button
-            className="blueBtn"
-            type="button"
-            onClick={() => history.push(`/meetups/${meetup.id}/edit`)}
-          >
+        <div>
+          <button className="blueBtn" type="button" onClick={handleEdit}>
             <MdEdit size={18} color="#fff" />
-            Editar
+            <span>Editar</span>
           </button>
           <button className="redBtn" type="button" onClick={handleCancel}>
             <MdDeleteForever size={18} color="#fff" />
-            Cancelar
+            <span>Cancelar</span>
           </button>
-        </span>
+        </div>
       </header>
       <img
         src={
